Account for manually set padding when scrolling.
Padding can be used to achieve better alignment. For example, it can be
used to align the first tab to the title of a Toolbar [as in the
Material design spec][1].

If android:paddingLeft or android:paddingRight are set in the XML, the
tab container is too wide and will scroll even if all of the tabs fit.

This commit changes the view to account for its padding when
calculating the container width. If app:pstsPaddingMiddle is true, the
view will behave as before.

[1]: https://www.google.com/design/spec/components/tabs.html#tabs-types-of-tabs